backport "init obs before vue" from slobs
https://github.com/stream-labs/desktop/pull/1959
